Run model training
