+++ src/app/model/login_hf.py
@@ -1,0 +1,3 @@
+from huggingface_hub import notebook_login
+
+notebook_login()
