+++ src/app/model/train_data.csv
@@ -1,0 +1,14 @@
+resources/srEdstva/train_1.mp3;СрЕдства
+resources/srEdstva/train_2.mp3;СрЕдства
+resources/srEdstva/train_3.mp3;СрЕдства
+resources/srEdstva/train_4.mp3;СрЕдства
+resources/srEdstva/train_5.mp3;СрЕдства
+resources/srEdstva/train_6.mp3;СрЕдства
+resources/srEdstva/train_7.mp3;СредствА
+resources/srEdstva/train_8.mp3;СредствА
+resources/srEdstva/train_9.mp3;СредствА
+resources/srEdstva/train_10.mp3;СредствА
+resources/srEdstva/train_11.mp3;СредствА
+resources/srEdstva/train_12.mp3;СредствА
+resources/srEdstva/train_14.wav;СредствА
+resources/srEdstva/train_15.mp3;СредствА
