--- src/app/training.py
@@ -1,74 +1,0 @@
-from transformers import WhisperFeatureExtractor, WhisperTokenizer, WhisperProcessor, WhisperForConditionalGeneration, Seq2SeqTrainingArguments, Seq2SeqTrainer
-
-from src.app.DataCollator import data_collator
-from src.app.dataset import common_voice
-from src.app.metrics import compute_metrics
-
-feature_extractor = WhisperFeatureExtractor.from_pretrained("openai/whisper-small")
-tokenizer = WhisperTokenizer.from_pretrained("openai/whisper-small", language="russian", task="transcribe")
-processor = WhisperProcessor.from_pretrained("openai/whisper-small", language="russian", task="transcribe")
-model = WhisperForConditionalGeneration.from_pretrained("openai/whisper-small")
-
-def prepare_dataset(batch):
-    # load and resample audio data from 48 to 16kHz
-    audio = batch["audio"]
-
-    # compute log-Mel input features from input audio array
-    batch["input_features"] = feature_extractor(audio["array"], sampling_rate=audio["sampling_rate"]).input_features[0]
-
-    # encode target text to label ids
-    batch["labels"] = tokenizer(batch["transalation"]).input_ids
-    return batch
-
-model.generation_config.language = "russian"
-model.generation_config.task = "transcribe"
-
-model.generation_config.forced_decoder_ids = None
-
-training_args = Seq2SeqTrainingArguments(
-    output_dir="./whisper-small-hi",  # change to a repo name of your choice
-    per_device_train_batch_size=16,
-    gradient_accumulation_steps=1,  # increase by 2x for every 2x decrease in batch size
-    learning_rate=1e-5,
-    warmup_steps=5,
-    max_steps=10,
-    gradient_checkpointing=True,
-    fp16=True,
-    evaluation_strategy="steps",
-    per_device_eval_batch_size=8,
-    predict_with_generate=True,
-    generation_max_length=225,
-    save_steps=10,
-    eval_steps=10,
-    logging_steps=25,
-    report_to=["tensorboard"],
-    load_best_model_at_end=True,
-    metric_for_best_model="wer",
-    greater_is_better=False,
-    push_to_hub=False,
-)
-
-trainer = Seq2SeqTrainer(
-    args=training_args,
-    model=model,
-    train_dataset=common_voice["train"],
-    eval_dataset=common_voice["test"],
-    data_collator=data_collator,
-    compute_metrics=compute_metrics,
-    tokenizer=processor.feature_extractor,
-)
-
-trainer.train()
-
-# check model after training
-# pipe = pipeline(
-#     "automatic-speech-recognition",
-#     model=model,
-#     tokenizer=tokenizer,
-#     feature_extractor=feature_extractor,
-#     torch_dtype=torch_dtype,
-#     device=device,
-# )
-#
-# result = pipe("Запись (21).mp3", generate_kwargs={"language": "russian"})
-# print(result["text"])
